+++ src/main/groovy/com/readytalk/gradle/js/NpmCachePlugin.groovy
@@ -1,10 +1,16 @@
 package com.readytalk.gradle.js
 
+import com.moowork.gradle.node.NodePlugin
+import com.moowork.gradle.node.task.NpmTask
+import com.moowork.gradle.node.task.SetupTask
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.Task
+
 import java.security.MessageDigest
 
 class NpmCachePlugin implements Plugin<Project> {
+  static final String NPM_INSTALL_WRAPPER = 'realNpmInstall'
   Project project
 
   void apply(Project project) {
@@ -12,18 +18,25 @@ class NpmCachePlugin implements Plugin<Project> {
     project.plugins.apply(JsPlugin)
 
     project.with {
-      tasks.replace('installGrunt').configure {}
+      tasks.replace('installGrunt')
+
+      Task npmTask = tasks.create(name: NPM_INSTALL_WRAPPER, type: NpmTask) {
+        args = ['install']
+      }
 
       tasks.replace('npmInstall').configure {
+        dependsOn tasks.findByName(SetupTask.NAME)
         inputs.file 'package.json'
         inputs.property 'NodeVersion', project.node.version
 
         def hashFile = file("${buildDir}/.package_json_hash")
         outputs.file hashFile
 
+        ext.cacheDir = new File("${System.properties['user.home']}/.node_modules_cache")
+
         doLast {
-          String hash = project.node.version + " - " + project.node.npmVersion + " - " + sha1(file('package.json'))
-          File f = new File("${System.properties['user.home']}/.node_modules_cache/${hash}.tar.gz")
+          String hash = project.node.version + "-" + project.node.npmVersion + "-" + sha1(project.file('package.json'))
+          File f = new File("${ext.cacheDir.absolutePath}/${hash}.tar.gz")
           if (f.exists()) {
             exec {
               executable 'rm'
@@ -47,10 +60,7 @@ class NpmCachePlugin implements Plugin<Project> {
               args '-rf', 'node_modules'
             }
 
-            def npm = tasks.create(name: 'real_npmInstall', type: NpmTask)
-
-            npm.args = ['install']
-            npm.execute()
+            npmTask.execute()
 
             f.parentFile.mkdirs()
 
@@ -68,7 +78,7 @@ class NpmCachePlugin implements Plugin<Project> {
     }
   }
 
-  private String sha1(File f) {
+  static String sha1(File f) {
     MessageDigest md = MessageDigest.getInstance("SHA-1");
     f.eachByte 4096, {bytes, size ->
       md.update(bytes, 0, size);