+++ src/test/groovy/com/readytalk/gradle/js/NpmCacheTest.groovy
@@ -1,7 +1,0 @@
-package com.readytalk.gradle.js
-
-import nebula.test.PluginProjectSpec
-
-class NpmCacheTest extends PluginProjectSpec {
-  String pluginName = 'com.readytalk.npmcache'
-}